move reflection fixture in fixture file
